# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: getEmployees
      tags: [employees]
      summary: List employees with current leave balances
      parameters:
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: Employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
  /me/employee:
    get:
      operationId: getMyEmployee
      tags: [employees]
      summary: Get the signed-in employee profile
      responses:
        "200":
          description: Employee profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "401":
          description: Authentication required
        "404":
          description: Employee profile not registered
    post:
      operationId: registerMyEmployee
      tags: [employees]
      summary: Register the signed-in employee profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeRegistrationInput"
      responses:
        "201":
          description: Registered employee profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "401":
          description: Authentication required
  /employees/{id}/balance:
    get:
      operationId: getEmployeeBalance
      tags: [employees]
      summary: Get an employee's leave balance
      parameters:
        - $ref: "#/components/parameters/EmployeeId"
      responses:
        "200":
          description: Leave balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveBalance"
        "404":
          description: Employee not found
...
  /leave-requests:
    get:
      operationId: getLeaveRequests
      tags: [leave-requests]
      summary: List leave requests
      parameters:
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Year"
        - $ref: "#/components/parameters/Month"
        - $ref: "#/components/parameters/EmployeeIdQuery"
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: createLeaveRequest
      tags: [leave-requests]
      summary: Submit an annual leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveRequestInput"
      responses:
        "201":
          description: Created leave request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
        "400":
          description: Invalid request
  /attendance/absences:
    post:
      operationId: createAbsence
      tags: [leave-requests]
      summary: Record an employee absence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbsenceInput"
      responses:
        "201":
          description: Recorded absence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
        "400":
          description: Invalid request
  /leave-requests/{id}:
    get:
      operationId: getLeaveRequest
      tags: [leave-requests]
      summary: Get a leave request
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Leave request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
        "404":
          description: Leave request not found
  /leave-requests/{id}/status:
    patch:
      operationId: updateLeaveRequestStatus
      tags: [leave-requests]
      summary: Approve or reject a leave request
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveRequestStatusInput"
      responses:
        "200":
          description: Updated leave request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
        "404":
          description: Leave request not found
components:
  parameters:
    EmployeeId:
      name: id
      in: path
      required: true
      schema:
        type: number
    RequestId:
      name: id
      in: path
      required: true
      schema:
        type: number
    Search:
      name: search
      in: query
      required: false
      schema:
        type: string
    Status:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/LeaveRequestStatus"
    Year:
      name: year
      in: query
      required: false
      schema:
        type: number
    Month:
      name: month
      in: query
      required: false
      schema:
        type: number
        minimum: 1
        maximum: 12
    EmployeeIdQuery:
      name: employeeId
      in: query
      required: false
      schema:
        type: number
  schemas:
...
    Employee:
      type: object
      required: [id, name, department, position, role, joinedAt, status, ordinaryHourlyWage, annualAllowance, usedDays, pendingDays, remainingDays]
      properties:
        id:
          type: number
        name:
          type: string
        department:
          type: string
        role:
          type: string
        joinedAt:
          type: string
          format: date
        annualAllowance:
          type: number
        usedDays:
          type: number
        pendingDays:
          type: number
        remainingDays:
          type: number
        companyName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        employeeNumber:
          type: ["string", "null"]
        position:
          type: string
        status:
          type: string
          enum: [active, resigned]
        resignedAt:
          type: ["string", "null"]
          format: date
        ordinaryHourlyWage:
          type: number
    EmployeeRegistrationInput:
      type: object
      required: [companyName, name, email]
      properties:
        companyName:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 3
...
    LeaveBalance:
      type: object
      required: [employeeId, annualAllowance, usedDays, pendingDays, remainingDays]
      properties:
        employeeId:
          type: number
        annualAllowance:
          type: number
        usedDays:
          type: number
        pendingDays:
          type: number
        remainingDays:
          type: number
    LeaveRequestStatus:
      type: string
      enum: [pending, approved, rejected, cancelled]
    LeaveType:
      type: string
      enum: [annual, half_day, quarter_day, early_leave, outing, wedding_funeral, paid_leave, sick_leave, substitute, absence, sabbatical]
...
    LeaveRequest:
      type: object
      required: [id, employeeId, employeeName, department, leaveType, startDate, endDate, days, reason, status, createdAt]
      properties:
        id:
          type: number
        employeeId:
          type: number
        employeeName:
          type: string
        department:
          type: string
        leaveType:
          $ref: "#/components/schemas/LeaveType"
        timeSlot:
          type: ["string", "null"]
          enum: [start, end, null]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        days:
          type: number
        reason:
          type: string
        status:
          $ref: "#/components/schemas/LeaveRequestStatus"
        createdAt:
          type: string
          format: date-time
        processedAt:
          type: ["string", "null"]
          format: date-time
        rejectionReason:
          type: ["string", "null"]
    LeaveRequestInput:
      type: object
      required: [employeeId, leaveType, startDate, endDate, days, reason]
      properties:
        employeeId:
          type: number
        leaveType:
          $ref: "#/components/schemas/LeaveType"
        timeSlot:
          type: ["string", "null"]
          enum: [start, end, null]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        days:
          type: number
          minimum: 0.25
        reason:
          type: string
          minLength: 1
    AbsenceInput:
      type: object
      required: [employeeId, startDate, endDate, days, reason]
      properties:
        employeeId:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        days:
          type: number
          minimum: 0.5
        reason:
          type: string
          minLength: 1
    LeaveRequestStatusInput:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/LeaveRequestStatus"
        rejectionReason:
          type: ["string", "null"]
...